simplifying the totals, based on subcomponent results
